pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "health-monitor-tg-bot"
version = "0.1.0"
description = "Telegram bot for tracking health measurements"
requires-python = ">=3.10"
dependencies = [
  "alembic>=1.16.5",
  "anyio==4.2.0",
  "certifi==2023.11.17",
  "h11==0.14.0",
  "httpcore==1.0.2",
  "httpx==0.25.2",
  "idna==3.6",
  "python-telegram-bot==20.7",
  "sniffio==1.3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pytest-asyncio>=0.21",
  "pytest-cov>=4.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
  "ignore::DeprecationWarning",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]



